+++ HetznerCloudApi/Object/Action/Get/Response.cs
@@ -2,12 +2,12 @@
 using Newtonsoft.Json;
 using System.Collections.Generic;
 
-namespace HetznerCloudApi.Object.Action
+namespace HetznerCloudApi.Object.Action.Get
 {
     public class Response
     {
         [JsonProperty("actions", NullValueHandling = NullValueHandling.Ignore)]
-        public List<Universal.Action> Actions { get; set; } = new List<Universal.Action>();
+        public List<Action> Actions { get; set; } = new List<Action>();
 
         [JsonProperty("meta", NullValueHandling = NullValueHandling.Ignore)]
         public Meta Meta { get; set; } = new Meta();