+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Version 0.0.5 (2023-11-02)
 - Global; The object properties have been changed to avoid conflicts with null values. This is done for compatibility with .NET Framework (This change is transparent to the user)
 - Volume; Complete
+- Volume Actions; Complete
 
 ## Version 0.0.4 (2023-11-02)
 - Datacenters; Complete